inlined parse_file_to_dom
this is required to prevent memory issues because of how we init the allocator with a pointer to the pool, which is also stored on the parser itself.  So by being inline we don't invalidate that pointer. I should probably just pull out the pool and only have the allocator on the script, then make a convenience macro/inline proc for setting it up with a pool